ml_backend/backend/requirements.txt
langchain
python-dotenv
langchain-google-vertexai
langchain-google-genai
fastapi
regex
uvicorn
python-multipart
pandas
tensorflow
wfdb
scikit-learn
faiss-cpu 
